requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0

# Data & ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6

# APIs & Network
stripe==5.4.0
requests==2.31.0

# Testing
pytest==7.4.0

# Utilities
python-dateutil==2.8.2
pydantic==1.10.12

